+++ mercata/ui/src/components/dashboard/AssetCard.tsx
@@ -1,4 +1,3 @@
-
 import { Link } from 'react-router-dom';
 import {
   Card,
@@ -13,20 +12,40 @@ interface AssetCardProps {
   symbol: string;
   price: string;
   deposit: string;
+  image?: string;
+  customDecimals: number;
 }
 
-const AssetCard = ({ id, name, symbol, price, deposit }: AssetCardProps) => {
+const AssetCard = ({ id, name, symbol, price, deposit, image, customDecimals }: AssetCardProps) => {
+  // Safely handle price and deposit as BigInt for formatUnits
+  let formattedPrice = "0.00";
+  let formattedDeposit = "0.00";
+  try {
+    formattedPrice = formatUnits(BigInt(price || "0"), customDecimals);
+  } catch {}
+  try {
+    formattedDeposit = formatUnits(BigInt(deposit || "0"), customDecimals);
+  } catch {}
+
   return (
     <Card className="h-full">
       <CardContent className="p-6">
         <div className="flex items-center justify-between mb-4">
           <div className="flex items-center space-x-3">
-            <div
-              className="w-10 h-10 rounded-full flex items-center justify-center text-white text-sm"
-              style={{ backgroundColor: "red" }}
-            >
-              {symbol?.slice(0, 2)}
-            </div>
+            {image ? (
+              <img
+                src={image}
+                alt={symbol}
+                className="w-10 h-10 rounded-full object-cover border"
+              />
+            ) : (
+              <div
+                className="w-10 h-10 rounded-full flex items-center justify-center text-white text-sm"
+                style={{ backgroundColor: "red" }}
+              >
+                {symbol?.slice(0, 2)}
+              </div>
+            )}
             <div>
               <h3 className="font-medium">{name}</h3>
               <p className="text-sm text-gray-500">{symbol}</p>
@@ -37,15 +56,11 @@ const AssetCard = ({ id, name, symbol, price, deposit }: AssetCardProps) => {
         <div className="space-y-2 mb-4">
           <div className="flex justify-between">
             <span className="text-sm text-gray-500">Price</span>
-              <span className="font-medium">{formatUnits(price.toLocaleString('fullwide', { useGrouping: false }),18)}</span>
+            <span className="font-medium">{formattedPrice}</span>
           </div>
           <div className="flex justify-between">
             <span className="text-sm text-gray-500">Asset Deposits</span>
-            <span className="font-medium">
-              {deposit
-                ? formatUnits(Number(deposit).toLocaleString('fullwide', { useGrouping: false }), 18)
-                : "0.00"}
-            </span>
+            <span className="font-medium">{formattedDeposit}</span>
           </div>
         </div>
