+++ mercata/ui/src/context/UserTokensContext.tsx
@@ -7,9 +7,12 @@ import isEqual from "lodash.isequal";
 type UserTokensContextType = {
   activeTokens: Token[];
   inactiveTokens: Token[];
+  allActiveTokens: Token[];
   loading: boolean;
+  allActiveLoading: boolean;
   error: string | null;
   fetchTokens: (userAddress: string, signal?: AbortSignal) => Promise<void>;
+  fetchAllActiveTokens: (signal?: AbortSignal) => Promise<void>;
 };
 
 const UserTokensContext = createContext<UserTokensContextType | undefined>(
@@ -21,7 +24,9 @@ export const UserTokensProvider: React.FC<{ children: React.ReactNode }> = ({
 }) => {
   const [activeTokens, setActiveTokens] = useState<Token[]>([]);
   const [inactiveTokens, setInactiveTokens] = useState<Token[]>([]);
+  const [allActiveTokens, setAllActiveTokens] = useState<Token[]>([]);
   const [loading, setLoading] = useState(false);
+  const [allActiveLoading, setAllActiveLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
 
   const fetchTokens = useCallback(async (userAddress: string, signal?: AbortSignal) => {
@@ -36,7 +41,7 @@ export const UserTokensProvider: React.FC<{ children: React.ReactNode }> = ({
       if (signal?.aborted) return;
 
       const allTokens = response.data || [];
-
+  
       const active = allTokens.filter((token: Token) => token.token.status === '2');
       const inactive = allTokens.filter((token: Token) => token.token.status !== '2');
 
@@ -62,15 +67,47 @@ export const UserTokensProvider: React.FC<{ children: React.ReactNode }> = ({
     }
   }, []);
 
+  // Modified to work like fetchTokens - storing in state
+  const fetchAllActiveTokens = useCallback(async (signal?: AbortSignal): Promise<void> => {
+    setAllActiveLoading(true);
+    setError(null);
+    try {
+      const response = await api.get(`/tokens?status=eq.2`, { signal });
+      // Only update state if not aborted and data has actually changed
+      if (signal?.aborted) return;
+      const tokens = response.data || [];
+      setAllActiveTokens(tokens);
+    } catch (err: any) {
+      if (
+        axios.isCancel?.(err) ||
+        err?.name === "CanceledError" ||
+        err?.code === "ERR_CANCELED" ||
+        err?.message === "canceled"
+      ) {
+        return;
+      }
+      console.error("Failed to fetch all active tokens:", err);
+      setError("Failed to fetch all active token data");
+      setAllActiveTokens([]);
+    } finally {
+      if (!signal?.aborted) {
+        setAllActiveLoading(false);
+      }
+    }
+  }, []);
+
   const contextValue = useMemo(
     () => ({
       activeTokens,
       inactiveTokens,
+      allActiveTokens,
       loading,
+      allActiveLoading,
       error,
       fetchTokens,
+      fetchAllActiveTokens,
     }),
-    [activeTokens, inactiveTokens, loading, error, fetchTokens]
+    [activeTokens, inactiveTokens, allActiveTokens, loading, allActiveLoading, error, fetchTokens, fetchAllActiveTokens]
   );
 
   return (
